+++ src/ui/dialog_scroll.rs
@@ -60,6 +60,7 @@ pub struct PlayerChoice(pub usize);
 ///     or update Text in existing dialogBox.
 ///   - ui::dialog_panel::update_player_scroll
 ///     - update choices displayed in the player scroll.
+#[derive(Event)]
 pub struct UpdateScrollEvent;
 
 /// # Note